feat: implement Order Line Quantity Update Feature (#6)
* feat: Create Order and OrderLine entities

This commit introduces the `Order` and `OrderLine` entities along with their associations and mappings. These entities are essential components for managing orders and order lines in the application.

* refactor: update BaseEntity class

In this commit, the BaseEntity class has been updated to include a setter for the 'id' field.

* feat: create order DTO classes

In this commit, the DTO classes OrderRequest, OrderResponse, OrderLineRequest, and OrderLineResponse have been created to facilitate data transfer between different parts of the application.

* feat: create OrderRepository and OrderService interfaces

In this commit, the OrderRepository interface and the OrderService interface have been created. OrderRepository provides methods for data access and manipulation related to orders, while OrderService defines service methods for order management.

* feat: create OrderSpecifications class

In this commit, the OrderSpecifications class has been created to provide specifications for querying orders based on product, customer, date from, and date to criteria. This class assists in building complex queries for order retrieval.

* feat: add getCustomerById method to CustomerService and its implementation

In this commit, the `getCustomerById` method has been added to the CustomerService interface and implemented in the CustomerServiceImpl class. This method allows for the retrieval of customer information by customer ID.

* feat: implement Order Service

This commit introduces the implementation of two key methods in the OrderServiceImpl class:

1. `createOrder`: This method creates orders based on the provided OrderRequest, ensuring that the customer exists and products are valid.

2. `getFilteredOrders`: This method retrieves filtered orders using specifications. It provides the ability to query and retrieve specific sets of orders based on criteria.

* feat: Implement OrderController

This commit introduces the OrderController class, which provides the following endpoints:

1. `createOrder`: This POST endpoint allows the creation of orders based on the provided OrderRequest. It ensures the request is validated and returns the created order as a response.

2. `searchOrders`: This GET endpoint enables searching for orders based on various criteria, such as product ID, customer ID, date range, etc. It constructs dynamic specifications based on the provided parameters and retrieves filtered orders.

The OrderController enhances the application's functionality by facilitating order creation and retrieval.

* docs: Add comments to OrderService interface

* refactor: improve OrderService readability

In this commit, the OrderServiceImpl class has been refactored to improve code readability and maintainability. The changes made include:

1. Removed unnecessary error handling for customer retrieval, as it is now handled in the CustomerService.getCustomerById method.
2. Created a private method, getProductMap, to improve code reuse and readability.
3. Extracted the creation of OrderLine objects into a separate method, createOrderLine, for better code organization.
4. Simplified the creation of the new order by directly setting its customer and order lines.
5. Removed unnecessary imports and logging statements to clean up the code.

* test: add OrderControllerTest and OrderServiceTest

* docs: improved documentation for the 'getProductMap', 'createOrderLine' methods in OrderServiceImpl

* refactor: Remove unnecessary @RequestBody and @Validated annotations

- Removed the @RequestBody and @Validated annotations from the createProduct and createCustomer methods in the service classes.

* test: update OrderServiceTest to Adapt to Modified Service Class

Refactored the OrderServiceTest class to accommodate changes in the OrderServiceImpl service class. Updated test cases to ensure the service functions correctly with the latest changes.

* feat: Add OrderNotFoundException and OrderLineNotFoundException

In this commit, we introduce two new exception classes:

1. `OrderNotFoundException`: This exception is thrown when an order is not found.
2. `OrderLineNotFoundException`: This exception is thrown when an order line is not found.

Both exceptions are annotated with `@ResponseStatus(HttpStatus.NOT_FOUND)` to indicate a 404 HTTP status code when thrown. They provide a constructor that allows specifying a custom error message for better error handling.

* feat: implement Order Line Quantity Update Feature

This commit introduces a new feature for updating the quantity of an order line within an existing order. Here's a summary of the changes made:

- Added a `updateOrderLineQuantity` method to the service layer, allowing to update the quantity of a specific order line in an order.

- Use exception `OrderLineNotFoundException`, to handle cases where the order line with the given ID is not found in the order.
- Added input validation using the `@RequestBody` annotation and a validation DTO (`OrderLineUpdateRequest`) to ensure the new quantity is valid.
- Updated the API endpoint to handle the quantity update request.
- Made improvements to the `Order` class to facilitate order line management.

The commit includes relevant error handling, validation, and documentation updates to ensure the proper functioning of this new feature.